+++ VideoApi/VideoApi.Services/Clients/Models/CreateConsultationRoomResponse.cs
@@ -6,13 +6,4 @@ public class CreateConsultationRoomResponse
 {
     [JsonPropertyName("room_label")] 
     public string RoomLabel { get; set; }
-
-    public override string ToString()
-    {
-        var sb = new StringBuilder();
-        sb.Append("class CreateConsultationRoomResponse {\n");
-        sb.Append("  RoomLabel: ").Append(RoomLabel).Append("\n");
-        sb.Append("}\n");
-        return sb.ToString();
-    }
 }
